+++ src/pages/OrderListPage/index.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import Loading from '../../components/Loading';
 import OrderItemListSections from '../../components/OrderList/OrderItemListSections';
 import ReactShoppingCartTemplate from '../../components/shared/ReactShoppingCartTemplate';
-import useFetch from '../../hooks/useFetch';
+import useFetch from '../../hooks/shared/useFetch';
 import { requestOrders } from '../../service/request/order';
 import { Order } from '../../types';
