+++ src/pages/ProductListPage/index.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import Loading from '../../components/Loading';
 import ProductList from '../../components/ProductList';
 import ReactShoppingCartTemplate from '../../components/shared/ReactShoppingCartTemplate';
-import useFetch from '../../hooks/useFetch';
+import useFetch from '../../hooks/shared/useFetch';
 import { requestProductList } from '../../service/request/productList';
 import { Product } from '../../types';
